+++ src/app/app-routing.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { NotFoundComponent } from './modules/core/components';
+import { CoreDemoComponent } from './modules/demo/components';
+
+const routes: Routes = [
+    {
+        path: 'demo',
+        component: CoreDemoComponent
+    },
+    {
+        path: '**',
+        component: NotFoundComponent
+    }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule {}