+++ src/app/app.module.ts
@@ -1,6 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { Router } from '@angular/router';
 
+import { AppRoutingModule } from './app-routing.module';
 import { ProductModule, CartModule, CoreModule, DemoModule } from './modules';
 import { AppComponent } from './app.component';
 
@@ -13,9 +15,14 @@ import { AppComponent } from './app.component';
     CoreModule,
     ProductModule,
     CartModule,
-    DemoModule
+    DemoModule,
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+    constructor(router: Router) {
+    console.log('Routes: ', JSON.stringify(router.config, undefined, 2));
+  }
+}