+++ src/app/modules/core/core.module.ts
@@ -1,11 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { NotFoundComponent } from './components';
+import { ProductService, CartService } from '../../services';
+
 @NgModule({
   imports: [
     CommonModule
   ],
-  exports: [],
-  declarations: []
+  exports: [NotFoundComponent],
+  declarations: [NotFoundComponent],
+  providers: [ProductService, CartService]
 })
 export class CoreModule { }