+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.4.1 - 2018-06-04
+### Added
+- Admin module
+- Login service (via local storage)
+- Login component
+- CanActivate guard for Admin module + CanActivateChild guard for its child navigation
+- Return URL support via query parameters
+
 ## 1.4.0 - 2018-05-31
 ### Added
 - Navigation menu